updated todolist-3
A decent spot to pull request this and ask for help. I've been able to get the remove item function working! Only problem is regardless of which button i click to remove a task, the top task is removed. I've had a look online for any help and general concensus is not to setState twice in a single function. I'm probably making the problem more difficult than it needs to be, but gonna sleep on it for now.