completed todolist-3
Followed your comments on the PR and I understand it completely. Sometimes when you've explained to me the correct way, I get how the whole thing works better because it naturally looks more logical.
I also added a couple extra touches such as when you hit return, it adds to the list, but also a line to stop a task being created if there is an empty string in the input field. A couple super small easy things but things that i'm proud of nontheless!