(almost!) completed part 4
So I actually almost got the whole thing working for part 4. I managed to turn the array of strings intio an array of objects and modify my code to work accordingly. something i was very proud of :smile:. The problem I have and that i'm too tired to spend any more time on is how to modify a value within an object with setState without changing any other key/value pairs in that object.