completed todolist-6.2
OK, it's at a point now where I'm happy to commit this for PR. I still think there's lots to be done styling wise, and I still need to give the user the option to choose between black or white text. I will do that by giving the user a drop down list with the options Black/White. Apart from that I would be happy to move on from here. Hope you like it!