completed todolist-7
Looks ok for now, and functions rather nicely. I don't think there are any major issues with this one since it's all just displaying the data in an alternate way, but I'm still pretty happy nonetheless.  I think the major work on this part for now is mostly styling based and I plan on creating a simple Show/Hide feature.
As for now, it's 6am and I'm going to bed.