completed todolist-8
Since Twitter was taking an age to reply to the API access request, I decided to try and add functionality to delete categories. I think it went quite well! The code ran first time, and after an initial bug where adding a category when there were none to begin with resulted in crashing, I altered the addCategory() function to allow for the category Ids to begin again at 1. Pretty happy!