significant progress on data hydration
I spent a while earlier today styling the program in CSS to look decent, then it was a matter of filling in the places where I could use the data response from the API. I feel it's a good place to check in with you now, as given I don't have access to Twitter's API yet, a lot of what I am doing is simply repeating the same thing over and over. I had a crack at putting links into the tweets and I came up with a solution but as soon as a tweet has multiple links, this stops working. I'd need to see how the API response is for more than one link first before having a crack at the logic. For now though, it works fine.